drop black.css in favour of a template
I used my python script to make a template of black.css with the base16
color names in place of the hexes. Then I modified the python script to
use this template, and dropped the step of replacing the original hexes
with the template ones. Now we can adjust the scheme by changing the
base16 names, and use the script to add the color hexes.